Using new settings API

+++ TcPlayer.Engine/Settings/SettingsFile.cs
@@ -1,14 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Linq;
-using System.Text;
 using System.Text.Json;
-using System.Threading.Tasks;
 
 namespace TcPlayer.Engine.Settings
 {
-    public sealed class SettingsFile
+    public sealed class SettingsFile: ISettingsFile
     {
         private readonly string _openFileName;
         private readonly Dictionary<Setting, string> _storage;
@@ -20,19 +17,21 @@ namespace TcPlayer.Engine.Settings
             _openFileName = fileName;
             if (File.Exists(_openFileName))
             {
-                var temp = JsonSerializer.Deserialize<Dictionary<Setting, string>>(File.ReadAllText(_openFileName));
+                var temp = JsonSerializer.Deserialize<Dictionary<string, Dictionary<string, string>>>(File.ReadAllText(_openFileName));
                 if (temp == null)
                     throw new InvalidOperationException();
 
-                _storage = temp;
-            }
-            else
-            {
-                _storage = new Dictionary<Setting, string>();
-                var empty = JsonSerializer.Serialize(_storage);
-                using (var f = File.CreateText(_openFileName))
+                foreach (var container in temp.Keys)
                 {
-                    f.Write(empty);
+                    foreach (var setting in temp[container])
+                    {
+                        _storage.Add(new Setting
+                        {
+                            Container = Guid.Parse(container),
+                            Key = setting.Key 
+                        },
+                        setting.Value);
+                    }
                 }
             }
             Settings = new SettingsProvider(_storage);
@@ -42,9 +41,31 @@ namespace TcPlayer.Engine.Settings
 
         public void Save()
         {
-            var newContents = JsonSerializer.Serialize(_storage);
+            var toSerialize = new Dictionary<string, Dictionary<string, string>>();
+
+            foreach (var item in _storage)
+            {
+                var containerId = item.Key.Container.ToString();
+                if (toSerialize.ContainsKey(containerId))
+                {
+                    toSerialize[containerId].Add(item.Key.Key, item.Value);
+                }
+                else
+                {
+                    toSerialize.Add(containerId, new Dictionary<string, string>());
+                    toSerialize[containerId].Add(item.Key.Key, item.Value);
+                }
+            }
+
+            var newContents = JsonSerializer.Serialize(toSerialize, new JsonSerializerOptions
+            {
+                WriteIndented = true,
+            });
             var oldfile = Path.ChangeExtension(_openFileName, ".old");
-            File.Move(_openFileName, oldfile);
+
+            if (File.Exists(_openFileName))
+                File.Move(_openFileName, oldfile);
+
             using (var f = File.CreateText(_openFileName))
             {
                 f.Write(newContents);
